packages: Replace DependenciesMapping with string descriptions
The only purpose of the DependenciesMapping enum is to differentiate the
various types of mappings stored inside a Package object, e.g. allowing
a Package object like this to be created:

  Package(
    "foo",
    {
      DependenciesMapping.LOCAL_ENV: {"bar", "baz"},
      DependenciesMapping.IDENTITY: {"foo"},
    }
  )

However, the way we are currently structuring our resolver stack, we
will never mix mappings from _different_ resolvers inside a single
Package object.

What we _do_ see, however, are Package objects coming from a single
resolver that nonetheless contain mappings that ultimately come from
different sources. This could be:

 - User-defined mappings spread across multiple files, where a single
   package gets imports contributed from different files.
 - Soon, we will be accepting multiple --pyenv options which could also
   allow a single package to appear in multiple environments, with
   different imports being contributed from different environments.

It would be very useful for debugging if the Package.mappings dict
allowed for more specific information about the source of a given
mapping.

By converting the keys in this dict from DependenciesMapping enum values
into free-form text strings, we can allow for this extra information,
consider for example the following Package objects in the future:

  Package(
    "foo",
    {
      "User-defined mapping from mappings.toml": {"foo"},
      "User-defined mapping from pyproject.toml": {"bar", "baz"},
    }
  )

or

  Package(
    "foo",
    {
      "Python env at .venv/lib/python3.7/site-packages": {"foo", "bar"},
      "Python env at __pypackages__/3.10/lib": {"foo", "baz"},
    }
  )

This commit removes the DependenciesMapping enum, and uses text strings
for the various types of mappings instead:

 - UserDefinedMapping: USER_DEFINED -> "User-defined mapping"
   (to be made more precise about the mapping source in a future commit)
 - LocalPackageResolver: LOCAL_ENV -> "Python env at {site_packages}"
   ("{site_packages}" is replaced by the dir containing the package)
 - TemporaryPipInstallResolver: LOCAL_ENV -> "Temporarily pip-installed"
 - IdentityMapping: IDENTITY -> "Identity mapping"