conftest: Return site_dir in addition to venv_dir from fake_venv()
The fake_venv() fixture is about to be put into service to attempt
further isolation of tests that use a default LocalPackageResolver
(i.e. one that looks at sys.path) from the virtualenv in which the
test happens to be running.

This will be done by prepending a fake_venv to sys.path, and for that
we need to know the site-packages directory created by the fake_venv.