package.UserDefinedMapping: Include source of mapping in Package objects
When accumulating a user-defined Package object from multiple files
(e.g. a [tool.fawltydeps.custom_mapping] section in pyproject.toml
together with a --custom-mapping-file=mapping.toml), include the
ultimate source of the mapping into the mapping description that is
stored into the Package object. This makes it easier to debug the source
of these user-defined meppings.

In other words, given a mapping.toml with:

  apache-airflow = ["airflow"]

and a pyproject.toml with:

  [tool.fawltydeps.custom_mapping]
  apache-airflow = ["foo", "bar"]

we used to generate this Package object:

  Package(
    "apache_airflow",
    {"User-defined mapping": {"airflow", "foo", "bar"}},
  )

but with this commit we instead get:

  Package(
    "apache_airflow",
    {
      "User-defined mapping from mapping.toml": {"airflow"},
      "User-defined mapping from settings": {"foo", "bar"},
    },
  )