conftest: Partially isolate the default LocalPackageResolver
LocalPackageResolver without a given pyenv_path defaults to looking up
packages in sys.path. This is problematic in tests, as we do not fully
control the Python environment in which the tests are run (e.g. consider
a developer running pytest directly in their dev environment where they
happen to have installed some packages that we don't expect).

Until now, we've skirted around the problem by adding a couple of TODO
comments near the relevant tests, but it's now time to solve this in a
better way. However, short of running the tests in a subprocess, we
cannot _fully_ control the sys.path. What we can do, is to use a
combination of our fake_venv fixture and pytest's
monkeypatch.syspath_prepend() function to insert an environment that we
_do_ control at the start of sys.path. This will allow us to place
entries into that fake_venv that we can then expect to be resolved
exactly as we expect in tests.

We introduce a new isolate_default_resolver fixture to apply this
partial isolation for a given dict of package-to-imports mappings.
At the same time we add a default_sys_path_env_for_tests dict that
encodes what we previously considered the minimal/default package
mappings that we expected (but could not guarantee) to be present in
the test environment.

With this, we can now call

  isolate_default_resolver(default_sys_path_env_for_tests)

from tests that previously relied on the surrounding test environment,
and their expectation is now guarateed by this new fake_venv inserted at
the start of sys.path.

Note that for tests that rely on packages NOT being installed into the
test environment (e.g. testing the fallback to the identity mapping), we
must still take care to use package names that we can ~guarantee never
to be present in the environment in which the tests are run. Thus, we
are only _partially_ isolated from the surrounding test environment.