Revamp how TemporaryPipInstallResolver reuses a cached venv
The cached venv that is currently only used by test_resolver should not
need to burden TemporaryPipInstallResolver with a direct dependency on
pytest. Instead, use a .cached_venv class variable to allow
test_resolver to customize the TemporaryPipInstallResolver class with a
reusable venv directory (from pytest's cache).

TemporaryPipInstallResolver only needs to know where this reusable venv
dir exists, and the details of interacting with the pytest cache can be
hidden away in test_resolver.

Also simplify the context managers used in TemporaryPipInstallResolver
for establishing a context where the given requirements are installed:
Split out a new installed_requirements() context manager that does venv
creation (if needed) and package installation, and leaves the existing
temp_installed_requirements() context manager to only concern itself
with creating the temporary directory (and delegate to
installed_requirements() for the rest).

Using a simple class member to control the cached_venv logic also allows
us to avoid depending on pytest-mock for patching
TemporaryPipInstallResolver in a more complex manner.